Fixes. Reorganize tests to be ensure to test correct assertions framework
