+++ test/Serilog.Sinks.InMemory.Assertions.Tests.Unit.AwesomeAssertions9/Serilog.Sinks.InMemory.Assertions.Tests.Unit.AwesomeAssertions9.csproj
@@ -5,6 +5,7 @@
 
     <IsPackable>false</IsPackable>
     <DefineConstants>$(DefineConstants);AWESOMEASSERTIONS_9</DefineConstants>
+    <RootNamespace>Serilog.Sinks.InMemory.AssertionsTests</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
