+++ test/Serilog.Sinks.InMemory.Assertions.Tests.Unit.FluentAssertions5/Serilog.Sinks.InMemory.Assertions.Tests.Unit.FluentAssertions5.csproj
@@ -5,6 +5,7 @@
 
     <IsPackable>false</IsPackable>
     <DefineConstants>$(DefineConstants);FLUENTASSERTIONS_5</DefineConstants>
+    <RootNamespace>Serilog.Sinks.InMemory.AssertionsTests</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
