+++ test/Serilog.Sinks.InMemory.Assertions.Tests.Unit.FluentAssertions7/Serilog.Sinks.InMemory.Assertions.Tests.Unit.FluentAssertions7.csproj
@@ -5,6 +5,8 @@
 
     <IsPackable>false</IsPackable>
     <DefineConstants>$(DefineConstants);FLUENTASSERTIONS_7</DefineConstants>
+    <RootNamespace>Serilog.Sinks.InMemory.AssertionsTests</RootNamespace>
+    
   </PropertyGroup>
 
   <ItemGroup>
