+++ test/Serilog.Sinks.InMemory.Assertions.Tests.Unit.FluentAssertions8/Serilog.Sinks.InMemory.Assertions.Tests.Unit.FluentAssertions8.csproj
@@ -5,6 +5,7 @@
 
     <IsPackable>false</IsPackable>
     <DefineConstants>$(DefineConstants);FLUENTASSERTIONS_8</DefineConstants>
+    <RootNamespace>Serilog.Sinks.InMemory.AssertionsTests</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
