+++ test/Serilog.Sinks.InMemory.Assertions.Tests.Unit.Shouldly4/WhenAssertingAndSInkIsWrittenTo.cs
@@ -1,7 +1,4 @@
-﻿using Xunit;
-using Serilog.Sinks.InMemory.Assertions;
-
-namespace Serilog.Sinks.InMemory.Assertions.Tests.Unit
+﻿namespace Serilog.Sinks.InMemory.AssertionsTests
 {
     public class WhenAssertingAndSInkIsWrittenTo
     {
