+++ test/Serilog.Sinks.InMemory.Assertions.Tests.Unit.Shouldly4/WhenAssertingLogEventHasLevel.cs
@@ -1,11 +1,4 @@
-using System;
-using Shouldly;
-using Serilog.Events;
-using Xunit;
-using Xunit.Sdk;
-using Serilog.Sinks.InMemory.Assertions;
-
-namespace Serilog.Sinks.InMemory.Assertions.Tests.Unit
+namespace Serilog.Sinks.InMemory.AssertionsTests
 {
     public class WhenAssertingLogEventHasLevel
     {
