+++ test/Serilog.Sinks.InMemory.Assertions.Tests.Unit.Shouldly4/WhenAssertingPropertyValuesOnMultipleMessages.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using Shouldly;
-using Xunit;
-using Serilog.Sinks.InMemory.Assertions;
-using Xunit.Sdk;
-
-namespace Serilog.Sinks.InMemory.Assertions.Tests.Unit
+﻿namespace Serilog.Sinks.InMemory.AssertionsTests
 {
     public class WhenAssertingPropertyValuesOnMultipleMessages
     {
