+++ test/Serilog.Sinks.InMemory.Assertions.Tests.Unit/WhenAssertingLogEventHasLevel.cs
@@ -1,91 +1,79 @@
-using System;
-#if AWESOMEASSERTIONS_9
-using AwesomeAssertions;
-#else
-using FluentAssertions;
-#endif
-using Serilog.Events;
-using Xunit;
-using Xunit.Sdk;
-using Serilog.Sinks.InMemory.Assertions;
+namespace Serilog.Sinks.InMemory.AssertionsTests;
 
-namespace Serilog.Sinks.InMemory.Assertions.Tests.Unit
+public class WhenAssertingLogEventHasLevel
 {
-    public class WhenAssertingLogEventHasLevel
+    private readonly ILogger _logger;
+
+    public WhenAssertingLogEventHasLevel()
     {
-        private readonly ILogger _logger;
+        _logger = new LoggerConfiguration()
+            .WriteTo.InMemory()
+            .CreateLogger();
+    }
 
-        public WhenAssertingLogEventHasLevel()
-        {
-            _logger = new LoggerConfiguration()
-                .WriteTo.InMemory()
-                .CreateLogger();
-        }
+    [Fact]
+    public void GivenInformationMessageIsLoggedAndAssertingWarning_WithLevelFails()
+    {
+        _logger.Information("Hello, world!");
 
-        [Fact]
-        public void GivenInformationMessageIsLoggedAndAssertingWarning_WithLevelFails()
-        {
-            _logger.Information("Hello, world!");
+        Action action = () => InMemorySink
+            .Instance
+            .Should()
+            .HaveMessage("Hello, world!")
+            .Appearing().Once()
+            .WithLevel(LogEventLevel.Warning);
 
-            Action action = () => InMemorySink
-                .Instance
-                .Should()
-                .HaveMessage("Hello, world!")
-                .Appearing().Once()
-                .WithLevel(LogEventLevel.Warning);
+        action
+            .Should()
+            .Throw<XunitException>()
+            .WithMessage("Expected message \"Hello, world!\" to have level \"Warning\", but it is \"Information\"");
+    }
 
-            action
-                .Should()
-                .Throw<XunitException>()
-                .WithMessage("Expected message \"Hello, world!\" to have level \"Warning\", but it is \"Information\"");
-        }
+    [Fact]
+    public void GivenInformationMessageIsLoggedAndAssertingInformation_WithLevelSucceeds()
+    {
+        _logger.Information("Hello, world!");
 
-        [Fact]
-        public void GivenInformationMessageIsLoggedAndAssertingInformation_WithLevelSucceeds()
-        {
-            _logger.Information("Hello, world!");
+        InMemorySink
+            .Instance
+            .Should()
+            .HaveMessage("Hello, world!")
+            .Appearing().Once()
+            .WithLevel(LogEventLevel.Information);
+    }
 
-            InMemorySink
-                .Instance
-                .Should()
-                .HaveMessage("Hello, world!")
-                .Appearing().Once()
-                .WithLevel(LogEventLevel.Information);
-        }
+    [Fact]
+    public void GivenMultipleInformationMessagesAndAssertingInformation_WithLevelSucceeds()
+    {
+        _logger.Information("Hello, world!");
+        _logger.Information("Hello, world!");
+        _logger.Information("Hello, world!");
 
-        [Fact]
-        public void GivenMultipleInformationMessagesAndAssertingInformation_WithLevelSucceeds()
-        {
-            _logger.Information("Hello, world!");
-            _logger.Information("Hello, world!");
-            _logger.Information("Hello, world!");
+        InMemorySink
+            .Instance
+            .Should()
+            .HaveMessage("Hello, world!")
+            .Appearing().Times(3)
+            .WithLevel(LogEventLevel.Information);
+    }
 
-            InMemorySink
-                .Instance
-                .Should()
-                .HaveMessage("Hello, world!")
-                .Appearing().Times(3)
-                .WithLevel(LogEventLevel.Information);
-        }
+    [Fact]
+    public void GivenMultipleWarningMessagesAndAssertingInformation_WithLevelFails()
+    {
+        _logger.Warning("Hello, world!");
+        _logger.Warning("Hello, world!");
+        _logger.Warning("Hello, world!");
 
-        [Fact]
-        public void GivenMultipleWarningMessagesAndAssertingInformation_WithLevelFails()
-        {
-            _logger.Warning("Hello, world!");
-            _logger.Warning("Hello, world!");
-            _logger.Warning("Hello, world!");
+        Action action = () => InMemorySink
+            .Instance
+            .Should()
+            .HaveMessage("Hello, world!")
+            .Appearing().Times(3)
+            .WithLevel(LogEventLevel.Information);
 
-            Action action = () => InMemorySink
-                .Instance
-                .Should()
-                .HaveMessage("Hello, world!")
-                .Appearing().Times(3)
-                .WithLevel(LogEventLevel.Information);
-
-            action
-                .Should()
-                .Throw<XunitException>()
-                .WithMessage("Expected instances of log message \"Hello, world!\" to have level \"Information\", but found 3 with level \"Warning\"");
-        }
+        action
+            .Should()
+            .Throw<XunitException>()
+            .WithMessage("Expected instances of log message \"Hello, world!\" to have level \"Information\", but found 3 with level \"Warning\"");
     }
 }
