+++ test/Serilog.Sinks.InMemory.Assertions.Tests.Unit/WhenAssertingLogEventsExist.cs
@@ -1,131 +1,120 @@
-using System;
-#if AWESOMEASSERTIONS_9
-using AwesomeAssertions;
-#else
-using FluentAssertions;
-#endif
-using Xunit;
-using Xunit.Sdk;
-using Serilog.Sinks.InMemory.Assertions;
+namespace Serilog.Sinks.InMemory.AssertionsTests;
 
-namespace Serilog.Sinks.InMemory.Assertions.Tests.Unit
+public class WhenAssertingLogEventsExist
 {
-    public class WhenAssertingLogEventsExist
+    private readonly ILogger _logger;
+
+    public WhenAssertingLogEventsExist()
     {
-        private readonly ILogger _logger;
+        _logger = new LoggerConfiguration()
+            .WriteTo.InMemory()
+            .CreateLogger();
+    }
 
-        public WhenAssertingLogEventsExist()
-        {
-            _logger = new LoggerConfiguration()
-                .WriteTo.InMemory()
-                .CreateLogger();
-        }
+    [Fact]
+    public void GivenNoLogMessage_HaveMessageFails()
+    {
+        Action action = () => InMemorySink.Instance
+            .Should()
+            .HaveMessage("Hello, World");
 
-        [Fact]
-        public void GivenNoLogMessage_HaveMessageFails()
-        {
-            Action action = () => InMemorySink.Instance
-                .Should()
-                .HaveMessage("Hello, World");
+        action
+            .Should()
+            .Throw<XunitException>()
+            .WithMessage("Expected message \"Hello, World\" to be logged");
+    }
 
-            action
-                .Should()
-                .Throw<XunitException>()
-                .WithMessage("Expected message \"Hello, World\" to be logged");
-        }
+    [Fact]
+    public void GivenSingleLogMessage_HaveMessagePasses()
+    {
+        _logger.Information("Hello, World");
 
-        [Fact]
-        public void GivenSingleLogMessage_HaveMessagePasses()
-        {
-            _logger.Information("Hello, World");
+        InMemorySink.Instance
+            .Should()
+            .HaveMessage("Hello, World");
+    }
 
-            InMemorySink.Instance
-                .Should()
-                .HaveMessage("Hello, World");
-        }
+    [Fact]
+    public void GivenTheSameLogMessageFourTimesAndAssertingOnlyOnce_HaveMessageFails()
+    {
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
 
-        [Fact]
-        public void GivenTheSameLogMessageFourTimesAndAssertingOnlyOnce_HaveMessageFails()
-        {
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
+        Action action = () => InMemorySink.Instance
+            .Should()
+            .HaveMessage("Hello, World")
+            .Appearing().Once();
 
-            Action action = () => InMemorySink.Instance
-                .Should()
-                .HaveMessage("Hello, World")
-                .Appearing().Once();
+        action
+            .Should()
+            .Throw<XunitException>()
+            .WithMessage("Expected message \"Hello, World\" to appear exactly once, but it was found 4 times");
+    }
 
-            action
-                .Should()
-                .Throw<XunitException>()
-                .WithMessage("Expected message \"Hello, World\" to appear exactly once, but it was found 4 times");
-        }
+    [Fact]
+    public void GivenTheSameLogMessageFourTimesAndAssertingFourTimes_HaveMessageSucceeds()
+    {
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
 
-        [Fact]
-        public void GivenTheSameLogMessageFourTimesAndAssertingFourTimes_HaveMessageSucceeds()
-        {
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
+        InMemorySink.Instance
+            .Should()
+            .HaveMessage("Hello, World")
+            .Appearing().Times(4);
+    }
 
-            InMemorySink.Instance
-                .Should()
-                .HaveMessage("Hello, World")
-                .Appearing().Times(4);
-        }
+    [Fact]
+    public void GivenTheSameLogMessageFourTimesAndAssertingFiveTimes_HaveMessageFails()
+    {
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
 
-        [Fact]
-        public void GivenTheSameLogMessageFourTimesAndAssertingFiveTimes_HaveMessageFails()
-        {
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
+        Action action = () => InMemorySink.Instance
+            .Should()
+            .HaveMessage("Hello, World")
+            .Appearing().Times(5);
 
-            Action action = () => InMemorySink.Instance
-                .Should()
-                .HaveMessage("Hello, World")
-                .Appearing().Times(5);
+        action
+            .Should()
+            .Throw<XunitException>()
+            .WithMessage("Expected message \"Hello, World\" to appear 5 times, but it was found 4 times");
+    }
 
-            action
-                .Should()
-                .Throw<XunitException>()
-                .WithMessage("Expected message \"Hello, World\" to appear 5 times, but it was found 4 times");
-        }
+    [Fact]
+    public void GivenMessageExistsAndTestingForAbsence_AssertionFails()
+    {
+        _logger.Information("Hello, World");
 
-        [Fact]
-        public void GivenMessageExistsAndTestingForAbsence_AssertionFails()
-        {
-            _logger.Information("Hello, World");
+        Action action = () => InMemorySink.Instance
+            .Should()
+            .NotHaveMessage("Hello, World");
 
-            Action action = () => InMemorySink.Instance
-                .Should()
-                .NotHaveMessage("Hello, World");
+        action
+            .Should()
+            .Throw<XunitException>()
+            .WithMessage("Expected message \"Hello, World\" not to be logged, but it was found once");
+    }
 
-            action
-                .Should()
-                .Throw<XunitException>()
-                .WithMessage("Expected message \"Hello, World\" not to be logged, but it was found once");
-        }
+    [Fact]
+    public void GivenMessageExistsMultipleTImesAndTestingForAbsence_AssertionFailsWithAppearanceCountInFailure()
+    {
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
+        _logger.Information("Hello, World");
 
-        [Fact]
-        public void GivenMessageExistsMultipleTImesAndTestingForAbsence_AssertionFailsWithAppearanceCountInFailure()
-        {
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
-            _logger.Information("Hello, World");
+        Action action = () => InMemorySink.Instance
+            .Should()
+            .NotHaveMessage("Hello, World");
 
-            Action action = () => InMemorySink.Instance
-                .Should()
-                .NotHaveMessage("Hello, World");
-
-            action
-                .Should()
-                .Throw<XunitException>()
-                .WithMessage("Expected message \"Hello, World\" not to be logged, but it was found 3 times");
-        }
+        action
+            .Should()
+            .Throw<XunitException>()
+            .WithMessage("Expected message \"Hello, World\" not to be logged, but it was found 3 times");
     }
 }
