+++ test/Serilog.Sinks.InMemory.Assertions.Tests.Unit/WhenAssertingMessageExistsThatContainsPattern.cs
@@ -1,87 +1,76 @@
-﻿using System;
-#if AWESOMEASSERTIONS_9
-using AwesomeAssertions;
-#else
-using FluentAssertions;
-#endif
-using Xunit;
-using Xunit.Sdk;
-using Serilog.Sinks.InMemory.Assertions;
+﻿namespace Serilog.Sinks.InMemory.AssertionsTests;
 
-namespace Serilog.Sinks.InMemory.Assertions.Tests.Unit
+public class WhenAssertingMessageExistsThatContainsPattern
 {
-    public class WhenAssertingMessageExistsThatContainsPattern
+    [Fact]
+    public void GivenThreeMessagesWithPattern_AssertionSucceeds()
     {
-        [Fact]
-        public void GivenThreeMessagesWithPattern_AssertionSucceeds()
-        {
-            _logger.Information("padding pattern padding");
-            _logger.Information("pattern padding");
-            _logger.Information("padding pattern");
+        _logger.Information("padding pattern padding");
+        _logger.Information("pattern padding");
+        _logger.Information("padding pattern");
 
-            InMemorySink.Instance
-                .Should()
-                .HaveMessage()
-                .Containing("pattern")
-                .Appearing().Times(3);
-        }
+        InMemorySink.Instance
+            .Should()
+            .HaveMessage()
+            .Containing("pattern")
+            .Appearing().Times(3);
+    }
 
-        [Fact]
-        public void GivenThreeMessagesNotMatchingPattern_AssertionFails()
-        {
-            _logger.Information("padding pattern padding");
-            _logger.Information("pattern padding");
-            _logger.Information("padding pattern");
+    [Fact]
+    public void GivenThreeMessagesNotMatchingPattern_AssertionFails()
+    {
+        _logger.Information("padding pattern padding");
+        _logger.Information("pattern padding");
+        _logger.Information("padding pattern");
 
-            Action action = () => InMemorySink.Instance
-                .Should()
-                .HaveMessage()
-                .Containing("NOT MATCHING")
-                .Appearing().Times(3);
+        Action action = () => InMemorySink.Instance
+            .Should()
+            .HaveMessage()
+            .Containing("NOT MATCHING")
+            .Appearing().Times(3);
 
-            action
-                .Should()
-                .Throw<XunitException>()
-                .WithMessage("Expected a message with pattern \"NOT MATCHING\" to be logged");
-        }
+        action
+            .Should()
+            .Throw<XunitException>()
+            .WithMessage("Expected a message with pattern \"NOT MATCHING\" to be logged");
+    }
 
-        [Fact]
-        public void GivenThreeMessagesAndAssertingAppearanceCount_AssertionSucceeds()
-        {
-            _logger.Information("foo");
-            _logger.Information("bar");
-            _logger.Information("baz");
+    [Fact]
+    public void GivenThreeMessagesAndAssertingAppearanceCount_AssertionSucceeds()
+    {
+        _logger.Information("foo");
+        _logger.Information("bar");
+        _logger.Information("baz");
 
-            InMemorySink.Instance
-                .Should()
-                .HaveMessage()
-                .Appearing().Times(3);
-        }
+        InMemorySink.Instance
+            .Should()
+            .HaveMessage()
+            .Appearing().Times(3);
+    }
 
-        [Fact]
-        public void GivenThreeMessagesAndAssertingNoMessagesAreLogged_AssertionFails()
-        {
-            _logger.Information("foo");
-            _logger.Information("bar");
-            _logger.Information("baz");
+    [Fact]
+    public void GivenThreeMessagesAndAssertingNoMessagesAreLogged_AssertionFails()
+    {
+        _logger.Information("foo");
+        _logger.Information("bar");
+        _logger.Information("baz");
 
-            Action action = () => InMemorySink.Instance
-                .Should()
-                .NotHaveMessage();
+        Action action = () => InMemorySink.Instance
+            .Should()
+            .NotHaveMessage();
 
-            action
-                .Should()
-                .Throw<XunitException>()
-                .WithMessage("Expected no messages to be logged, but found 3 messages");
-        }
+        action
+            .Should()
+            .Throw<XunitException>()
+            .WithMessage("Expected no messages to be logged, but found 3 messages");
+    }
 
-        private readonly ILogger _logger;
+    private readonly ILogger _logger;
 
-        public WhenAssertingMessageExistsThatContainsPattern()
-        {
-            _logger = new LoggerConfiguration()
-                .WriteTo.InMemory()
-                .CreateLogger();
-        }
+    public WhenAssertingMessageExistsThatContainsPattern()
+    {
+        _logger = new LoggerConfiguration()
+            .WriteTo.InMemory()
+            .CreateLogger();
     }
 }
